Move stuff to autocompletion sub-package
