+++ feature/metis/conversation/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/ui/reply/MarkdownStyle.kt
@@ -1,0 +1,17 @@
+package de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply
+
+enum class MarkdownStyle(val startTag: String, val endTag: String) {
+    Bold("**", "**"),
+    Italic("*", "*"),
+    Underline("<ins>", "</ins>"),
+    Strikethrough("~~", "~~"),
+    InlineCode("`", "`"),
+    CodeBlock("```", "```"),
+    Blockquote("> ", ""),
+    OrderedList("1. ", ""),
+    UnorderedList("- ", ""),
+    UserMention("@", ""),
+    ChannelMention("#", ""),
+    LectureMention("#", ""),
+    ExerciseMention("#", "")
+}
