+++ feature/metis/conversation/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/ui/reply/autocomplete/AutoCompleteCategory.kt
@@ -1,4 +1,4 @@
-package de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply
+package de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.reply.autocomplete
 
 import androidx.annotation.StringRes
